# 05_WEBABLE_API_OpenAPI.yaml
openapi: 3.1.0
info:
  title: WEBABLE Public API
  version: 0.1.0
  description: |
    WEBABLE SaaS website builder API draft. 실제 구현 전 인증, 권한, idempotency, rate limit, audit logging 정책을 확정해야 한다.
servers:
  - url: https://api.webable.example.com
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteId:
      name: siteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AIJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [queued, running, needs_approval, applied, failed, cancelled]
        inputScope:
          type: object
        outputDiff:
          type: object
        cost:
          type: number
...
paths:
...
  /v1/sites/{siteId}/publish:
    post:
      summary: Publish site snapshot
      tags: [Publishing]
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        '202':
          description: Publish job accepted
...
  /v1/ai/site-generate:
    post:
      summary: Generate a site draft from onboarding prompt
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orgId, prompt]
              properties:
                orgId:
                  type: string
                  format: uuid
                prompt:
                  type: string
                industry:
                  type: string
                tone:
                  type: string
                targetAudience:
                  type: string
      responses:
        '202':
          description: AI job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIJob'
...